fixes and user experience update
included "product added to cart" message, filtered cart by session using a unique user id created by session, fixed cart update button and delete buttons and added some styling dor the added to cart message

+++ MCPharm/Order received.php
@@ -13,16 +13,18 @@ $db_port = $config['DB_PORT'];
 
 $conn=new mysqli($db_host, $db_user, $db_password, $db_name, $db_port);
 
+$user=$_SESSION['user_id'];
+
  $details="SELECT * FROM receipts";
  $run = mysqli_query($conn,$details);
- $item="SELECT p.Product_id, p.Product_Name,p.Brief_Description, p.Price,p.Product_Description,p.Image_path, c.Quantity 
+ $item="SELECT p.Product_id, p.Product_Name,p.Brief_Description, p.Price,p.Product_Description,p.Image_path, c.Quantity
  FROM products p 
  LEFT JOIN cart c ON p.Product_id = c.Product_id
- WHERE c.Quantity>=1;";
+ WHERE c.Quantity>=1 AND c.user_id=$user;";
  $result = mysqli_query($conn,$item);
 
 if($_SERVER["REQUEST_METHOD"]=="POST"){
-    $clear = "DELETE FROM receipts";
+    $clear = "DELETE FROM receipts WHERE user_id=$user";
     $reset = mysqli_query($conn,$clear);
     header("Location:homepage.html");
       exit();
@@ -37,7 +39,6 @@ if($_SERVER["REQUEST_METHOD"]=="POST"){
         <meta name="description" content="MCPharm is your trusted online pharmacy offering medications, wellness products, supplements, dental care, and cosmetics with fast delivery and personalized support.">
         <meta name="keywords" content="online pharmacy, medications, supplements, dental health, cosmetics, wellness products, MCPharm">
         <meta name="author" content="MCPharm">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link rel="stylesheet" href="style.css">
     </head>
